+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support deprecrated align attribute for block #171
 - Fix parsing of style attribute with a key with no value
 - Improve parsing of style attribute to avoid an extra call to HtmlDecode
+- Extend support of nested list for non-W3C compliant html #173
 
 ## 3.2.1