+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Fix parsing of style attribute with a key with no value
 - Improve parsing of style attribute to avoid an extra call to HtmlDecode
 - Extend support of nested list for non-W3C compliant html #173
+- Change way to apply table 100% width
+- Allow to apply percentage widths cells
 
 ## 3.2.1